k8s: Rename controller/ -> controllers/
