+++ provisioner/k8s/controllers/bundle_controller.go
@@ -1,4 +1,4 @@
-package controller
+package controllers
 
 import (
 	"context"
